store: Move primary-only functionality into a separate module
